# src/main/java/ar/edu/utn/frc/tup/lciii/dtos/common/CountryRequestDTO.java
package ar.edu.utn.frc.tup.lciii.dtos.common;

import lombok.*;

@Getter
@Setter
@Builder
@AllArgsConstructor
@NoArgsConstructor
public class CountryRequestDTO {
    private int amountOfCountryToSave;
}